fix: using jsonSchema def, clean up code, rename pages directory
