+++ src/applications/burial-poc-v6/app-entry.jsx
@@ -4,11 +4,9 @@ import './sass/burial-poc-v6.scss';
 import startApp from 'platform/startup/router';
 
 import routes from './routes';
-// import reducer from './reducers';
 import manifest from './manifest.json';
 
 startApp({
   url: manifest.rootUrl,
-  // reducer,
   routes,
 });
