--- src/applications/burial-poc-v6/config/form.js
@@ -1,55 +1,0 @@
-// // import fullSchema from 'vets-json-schema/dist/21P-530-schema.json';
-
-// import manifest from '../manifest.json';
-
-// import IntroductionPage from '../containers/IntroductionPage';
-// import ConfirmationPage from '../containers/ConfirmationPage';
-
-// // const { } = fullSchema.properties;
-
-// // const { } = fullSchema.definitions;
-
-// const formConfig = {
-//   rootUrl: manifest.rootUrl,
-//   urlPrefix: '/',
-//   // submitUrl: '/v0/api',
-//   submit: () =>
-//     Promise.resolve({ attributes: { confirmationNumber: '123123123' } }),
-//   trackingPrefix: 'burial-poc-v6-',
-//   introduction: IntroductionPage,
-//   confirmation: ConfirmationPage,
-//   formId: '21P-530',
-//   saveInProgress: {
-//     // messages: {
-//     //   inProgress: 'Your benefits application (21P-530) is in progress.',
-//     //   expired: 'Your saved benefits application (21P-530) has expired. If you want to apply for benefits, please start a new application.',
-//     //   saved: 'Your benefits application has been saved.',
-//     // },
-//   },
-//   version: 0,
-//   prefillEnabled: true,
-//   savedFormMessages: {
-//     notFound: 'Please start over to apply for benefits.',
-//     noAuth: 'Please sign in again to continue your application for benefits.',
-//   },
-//   title: 'burial-poc-v6',
-//   defaultDefinitions: {},
-//   chapters: {
-//     chapter1: {
-//       title: 'Chapter 1',
-//       pages: {
-//         page1: {
-//           path: 'first-page',
-//           title: 'First Page',
-//           uiSchema: {},
-//           schema: {
-//             type: 'object',
-//             properties: {},
-//           },
-//         },
-//       },
-//     },
-//   },
-// };
-
-// export default formConfig;
