+++ src/applications/burial-poc-v6/manifest.json
@@ -3,5 +3,6 @@
   "entryFile": "./app-entry.jsx",
   "entryName": "burial-poc-v6",
   "rootUrl": "/burial-poc-v6",
-  "productId": "3f7db617-af4b-4d7f-b30f-f288d5756668"
+  "productId": "3f7db617-af4b-4d7f-b30f-f288d5756668",
+  "production": false
 }
