+++ src/applications/burial-poc-v6/pages/BenefitsSelection.jsx
