+++ src/applications/burial-poc-v6/pages/BurialAllowance.jsx
