+++ src/applications/burial-poc-v6/pages/BurialIntroduction.jsx
