+++ src/applications/burial-poc-v6/pages/ClaimantInformation.jsx
