--- src/applications/burial-poc-v6/reducers/index.js
@@ -1,6 +1,0 @@
-import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
-import formConfig from '../config/form';
-
-export default {
-  form: createSaveInProgressFormReducer(formConfig),
-};
