+++ src/applications/burial-poc-v6/utils.js
@@ -31,14 +31,3 @@ export const stringToBoolean = value => {
 export const getRadioLabel = value => {
   return RadioLabels[value];
 };
-
-/**
- * A function to transform a dateString in the form yyyy-mm-dd into a javascript date object. This is necessary because
- * parsing the date string will result in a date one day before the expected value
- *
- * @param dateString
- */
-export const parseDate = dateString => {
-  const [yyyy, mm, dd] = dateString.split('-').map(str => parseInt(str, 10));
-  return new Date(yyyy, mm - 1, dd);
-};
